Use findObjectsForSrc to watch the referenced topology
With the goal of deduplicating the current code as much as possible and
help spreading the pattern across the operators, this patch adds the
topologyRef dynamic field to the existing glanceAPIWatchList. This
allows to both index the field (and use fieldSelector when the GlanceAPI
CR is retrieved), and remove the topologyFn/tpFn functions.
The watcher is now based on GenerationChangedPredicate, which only
triggers a reconcile when .Spec for the referenced topology changes.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>